refcator: gird,list cell 각각 생성
